Added the background sound only to level1
Im not sure how to add the sound to all levels and maybe loop it?
The main menu  can be opened in the UI folder, I dont know how to make the game to always starts in the Main Menu and also I forgot to go back (for example, when you click on the Settings button. Sorry if I explain this bad, let me know if you need anything!